test: cover email subsetting section (bbox/polygon split, rendering, date handling)
